backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.0
pydantic>=2.6.1
pydantic-settings>=2.1.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis==5.0.1
requests==2.31.0
lark-oapi>=1.5.0 # 飞书官方 Python SDK
httpx>=0.27.0
python-dotenv==1.0.1
PyMuPDF==1.25.2
pdf2zh==1.9.11
python-docx>=1.1.0
python-pptx>=0.6.23
openpyxl>=3.1.2
apscheduler>=3.10.4
feedparser>=6.0.10
readability-lxml>=0.8.1
beautifulsoup4>=4.12.3
lxml>=5.2.2
passlib[bcrypt]>=1.7.4
bcrypt<4.0.0
PyJWT>=2.8.0
python-multipart>=0.0.9
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
python-multipart>=0.0.9
alembic>=1.13.1
